move CancelTask into separate class
